+++ DND/DND/DND/include/DNDGUISwitchSprite2.h
@@ -28,7 +28,7 @@ namespace DND
 
 		SwitchSprite2* Clone();
 
-		void SetOrder(int order);
+		void SetOrder(float order);
 	private:
 		//子类控件需要重载
 		virtual bool _is_pickup() override;
